requirements.txt
asgiref==3.6.0
attrs==23.1.0
cffi==1.15.1
cryptography==40.0.2
Django==4.2.1
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
inflection==0.5.1
jsonschema==4.17.3
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.6.0
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
uritemplate==4.1.1
